chore: cleanup worker stuff from memory storage
